+++ src/main/java/insane96mcp/iguanatweaksreborn/module/misc/Misc.java
@@ -14,6 +14,7 @@ public class Misc extends Module {
 	public TempSpawner tempSpawner;
 	public VillagerNerf villagerNerf;
 	public Nerf nerf;
+	public Beacon beacon;
 
 	public Misc() {
 		super(Config.builder);
@@ -24,6 +25,7 @@ public class Misc extends Module {
 		tempSpawner = new TempSpawner(this);
 		villagerNerf = new VillagerNerf(this);
 		nerf = new Nerf(this);
+		beacon = new Beacon(this);
 		Config.builder.pop();
 	}
 
@@ -36,5 +38,6 @@ public class Misc extends Module {
 		tempSpawner.loadConfig();
 		villagerNerf.loadConfig();
 		nerf.loadConfig();
+		beacon.loadConfig();
 	}
 }